Remove Authentication Provider Implementation
Summary:
The OpenCMIS implementation has an AuthenticationProvider
implementation which is used to authenticate against the CMIS Server.
In the PHP Implementation we use Guzzle for the HTTP communication
which already supports authentication. So the authentication provider
is not required and should be deleted to reduce complexity.

More infos about Guzzle and authentication:
http://docs.guzzlephp.org/en/latest/clients.html#auth

Reviewers: claus.due, #forgetit-dev

Reviewed By: claus.due, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D359