Check if response exists before using it while throwing exception in AbstractBrowserBinding
Summary:
Sometimes there is no responsive available when doing a get request. For example
when the given binding url does not exists or is not reachable. In this case there empty
data should be passed instead of trying to access the non existing response which ends in
a php fatal error.

Test Plan:
Configure an invalid Browser binding endpoint and try to make a request.
Before this patch this ends in a php fatal. With this patch it should throw a CmisRuntimeException

Reviewers: dkd-ebert, #forgetit-dev

Reviewed By: dkd-ebert, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D541