Do not mock exception to fix hhvm error
Summary:
The unit tests currently fail for hhvm. See e.g.
https://travis-ci.org/dkd/php-cmis-client/jobs/47913193

This is because of a "missing feature" in phpunit. See
https://github.com/sebastianbergmann/phpunit-mock-objects/issues/207

Special methods that are added by hhvm to an exception are mocked by
php unit which ends in an error.

The exception should not be mocked to fix that problem.

Test Plan:
Run unit tests with hhvm. I use docker for that - if you have
questions than ask me.

Reviewers: claus.due, #forgetit-dev

Reviewed By: claus.due, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D673