Fix bugs introduced in D804 in Folder and Document
Summary:
In D804 a bug has been introduced at places where
`getNewlyCreatedObject` is used. This method returns null if
no $context is given which should then end up in returning null of
the calling function instead of throwing an exception and using a
non existend object in the exception handling.

Reviewers: dkd-ebert, #forgetit-dev

Reviewed By: dkd-ebert, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D807