Use `cmis:name` property as filename if stream uri misses file extension
Summary:
To send a file with a post request to the CMIS Server the guzzle
library is used. Guzzle gets the mime-type, that is used to transfer the file,
from the filename / stream. This could sometimes fail. For example if a file
is uploaded with php it gets a temporary name like `/tmp/f3jf92d9` and the
file is then tranferred as `text/plain` because the mime type could not be
detected by guzzle. To work around this problem the `cmis:name` property is
now used as fallback as this property should contain the real filename.

Guzzle Issue: https://github.com/guzzle/guzzle/issues/571

Test Plan:
Use the createDocument function to create a document with a filestream.
The file that is used should be for example a jpg image but the filename must
not contain a file extension. Now try to upload this image to the CMIS Server.

Expected result: The image should be uploaded and got the correct my type.
Old result: The image was uploaded as `text/plain` and could therefore not be
displayed in the CMIS Server.

Reviewers: claus.due, dkd-ebert, #forgetit-dev

Reviewed By: dkd-ebert, #forgetit-dev

Differential Revision: https://phabricator.dkd.de/D1175