Implement RepositoryService::createType functionality for browser binding
Summary:
Implement the createType functionality for the browser binding.
Added a new `TypeConverterInterface` which is used to convert object types to
a simple type representation used for the browser binding.

Test Plan:
As Alfresco does not support "createType" this can not be tested
with Alfresco as CMIS server.
The implementation has only been validated by the generated POST request.

Reviewers: dkd-ebert, #[t]_forgetit-dev

Reviewed By: dkd-ebert, #[t]_forgetit-dev

Differential Revision: https://phabricator.dkd.de/D1347