Use parent exception message if no message is defined in AbstractBrowserBindingService
Summary:
If a exception is thrown in the HTTP Invoker it is passed to the
`convertStatusCode` function where the body is checked for a json repsonse that
contains a error message.
If no message is found but a parent (Guzzle) exception is given the message of this
exception should be used in the thrown exception.

Reviewers: #[t]_forgetit-dev, dkd-ebert

Reviewed By: #[t]_forgetit-dev, dkd-ebert

Differential Revision: https://phabricator.dkd.de/D1290