Allow properties to not have an id
Summary:
In JSON Converter an exception was thrown when converting a property
if the `id` or `queryName` where null. But it should be only thrown if one of
these is null as A Property must always have an ID except if it used in a
query result. In a query result a Property should have an ID and must have a
query name.

This is now again in sync with the OpenCMIS Java implementation.

Reviewers: #[t]_forgetit-dev, claus.due

Reviewed By: #[t]_forgetit-dev, claus.due

Differential Revision: https://phabricator.dkd.de/D1330