[BUGFIX] Alias "choices" and "choice" on AbstractPropertyDefinition
Responses received from at least Alfresco will contain the "choice" wording as property name and this causes the "choices" property to not be set and triggers an Exception.

Occurs since renaming said property in the PHP library. Alias added to prevent failure while still allowing the property to be addressed as "choices" in PHP. Noted added in PHPDOC.