[BUGFIX] Use setContentStream for newly uploaded files
Due to the way posting file bodies works when combined with a query array, files with unicode names would receive an incorrect (latin1 variant) file name.

In order to avoid this happening, the file upload itself now gets performed in a separate request which posts only the file. Doing this lets the original createDocument call set the correct title first and it does not get a bad encoding after setting the file contents.

It is possible there is a more appropriate way to fix this (e.g. in PostFile settings) but I have not been able to find it. It seems to be caused by some detection on the CMIS server side and I've found no request flags that would seem to control it. Headers of the posted request appear to be the same when using either strategy.

This patch solves the issue for now and can easily be reverted / improved if a better way comes up.

Close: #22