Introduce read/post JSON returning wrappers
Methods currently call read() or post() which returns a
Response - however, almost all methods do not care
about the response but always want JSON data.

Previously, Guzzle allowed chaing a ->json call on a
Response to get decoded JSON as array but this is no
longer possible, so these methods are introduced to
again cut down the code duplication in Response parsing.