Fix some more undefined index errors
There are some places where the usage of an array cast in
combination with the null coalescing operator ends in an undefined
index error and the actual cast is not executed at all.
Fix those places by adding some braces around the null coalescing
check before casting the end result to an array.

Not sure if the array cast is required at all because the passed
value must be an array and casting it could lead to an unexpected
result in case of providing an invalid value. But that is a different
issue.